conformer/requirements.txt
mindspore_gpu~=1.9.0
numpy~=1.21.6
onnxruntime~=1.13.1
pandas~=1.3.5
Pillow~=9.4.0
PyYAML~=6.0
tqdm~=4.64.1
